Initial commit of challenge
This commit represents the changes necessary to setup this project
locally. This includes the following

- The `.ruby-version` was set to version 2.2.2 while the `Gemfile` ruby
version was set to 2.3.0. Since 2.2.2 uses an insecure version of
openssl the `.ruby-version` was updated.

- Ruby version 2.3.0 allows the use of Bundler 2.0 with very minimal
changes to the project's current gem versions. Only minitest and rake
needed to be updated.

- `database.yml` was changed to `database.yml.example` so that it could
still be shared as a template. And `config/database.yml` was added to
the `.gitignore` file so that the local database setup doesn't get added
to the repository