Added ability for users to create new pay account
In order for users to have the ability to create new pay accounts, the
following was added.

- Updates to the `create` and `new` pay_acount_controller actions. The
new controller action supplies an empty pay_account model for use in
associated view's form. The create action takes a filtered version of
the parameters returned by that form (known as "strong params"), uses
them to create a new pay_account object, and saves that object to the
database. It then redirects back to the index action so the user can
view their new pay_account with all the others

- Creation of a pay_account `new` view along with some minor css to size
it properly.

- Updates to the pay_account `index` view to link the create button to
the new view.